Don't create group if it spreads over other boxes
This can happen typically if two boxes are similar in shape and color
but other boxes are in between them. Ideally in such case any box
overlapping with the new group should be matched for inclusion and only
if it couldn't be included the merge process should be stopped.

In this first take of the approach, we are simply not permitting the
merge at all, no inclusion is considered.